Feature/stripe integration (#12)
* FEAT: add stripe gem dependency

* FEAT: add stripe keys and configurations

* FEAT: add stripe id to users

* FIX: database structure for torrents and users (adding stripe)

* FIX: logic for validating expired users

* FIX: quota testing with the db new default

* FEAT: add stripe and expires validations on user model w/tests

* CHORE: add stripe mock test dependency

* FEAT: user_creation service to create user and stripe account w/sub.

The subscription is to a basic plan that has 3 days trial

* FEAT: hook the UserCreation service to sign_up action

* FIX: add leaner comparison of expires_at

* FEAT: font-awesome dependency

* STYLE: fix rubocop warning

* FEAT: add the user expires_at to the status quota

* STYLE: fix rubocop validations

* FEAT: decorate user#expires_at w/test

* FIX: add jquery ujs to support link_to method: :delete

* FIX: better styling for login and control buttons

* FEAT: add user info partial styling

* FEAT: add dashboard link to navbar

* FEAT: add the views for the user info (WIP)

* FEAT: add the route and controller to user

* FEAT: hide e***l characters w/test

* FEAT: add subscription_id to user model to handle subscriptions

* FEAT: delegate_all unimplemented methods to user model

* FEAT: add the subscription scss files and links

* FEAT: add the link to the user subscription and use hidden helper

* FIX: email_hidden helper and subscribed? decorator w/test

* STYLE: remove commented unused code

* FEAT: add valid_subscription? method to user model

* REFACTOR: use user#valid_subscription? w/test

* FEAT: implement unsubscribe action

* REFACTOR: user creation w/test

* FIX: user model to always validate stripe fields

* FIX: user Fabricator

* CHORE: I18n for subscription responses

* FEAT: routes for subscription management

* FEAT: add SubscriptionError class

* REFACTOR: more readable code

* FEAT: add canceled_at and fix old Migration expires_at

* FIX: remove canceled_at

* FEAT: add UserPolicy for controlling subscriptions

* TEST: UserPolicy

* FEAT: add the subscription creation services

* TEST: SubscriptionCreation service

* FEAT: add SubscriptionCancellation service w/test

* FEAT: add the partial view of the subscriptions

* REFACTOR: subscribe/unsubscribe actions in the UsersController

* REFACTOR: remove dead code

* FIX: remove wrong logging or error

* FIX: UserCreation not to expose stripe errors

* FIX: validation of customer id on update

to avoid failing to save our record in the UserCreation Service before
we hit stripe API to register the customer and get his customer_id

this is done to avoid exposing the errors of stripe and it's email
validations to our front end

+++ db/migrate/20160723150458_create_torrents.rb
@@ -3,7 +3,7 @@ class CreateTorrents < ActiveRecord::Migration[5.0]
     create_table :torrents, id: :uuid do |t|
       t.string :name
       t.integer :transmission_id
-      t.integer :size
+      t.bigint :size
       t.string :checksum
       t.uuid :user_id, index: true
       t.timestamps
